Made the progress bar NICE
User can click a collapse button to chane the deadline of their project. This will also inform them of invalid dates and will update the prgress bar with the amount of days left. Will need to write the days left to memory to make comparison as the page loads to update the bar accordingly.