Made the chat a chat
Added functionality to the chat that allows users to send messages and the chat will update the log and clear the text input box. It's kinda jank rn but we can pretty it up later. Will also add functionality to add cards to the chat to add ideas into the chat.